[android] Fix build
