+++ app_manager/appmanager/src/main/java/app/editors/manager/managers/utils/NotificationUtils.java
@@ -1,180 +1,216 @@
+/*
+ * Created by Michael Efremov on 26.10.20 13:23
+ */
+
 package app.editors.manager.managers.utils;
 
-import android.app.DownloadManager;
 import android.app.Notification;
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
-import android.app.Service;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Build;
-import android.os.Environment;
-import android.os.StatFs;
+import android.os.Bundle;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.core.app.NotificationCompat;
-
-import java.io.File;
 
 import app.editors.manager.R;
 import app.editors.manager.app.App;
-import app.editors.manager.managers.services.DownloadService;
-import app.editors.manager.managers.services.UploadService;
+import app.editors.manager.managers.receivers.DownloadReceiver;
+import app.editors.manager.managers.receivers.UploadReceiver;
+import app.editors.manager.ui.activities.main.MainActivity;
 import lib.toolkit.base.managers.utils.ActivitiesUtils;
 
 public class NotificationUtils {
 
-    private static final int CANCELED_GROUP_ID = 1;
-    private static final String CANCELED_GROUP_KEY = "CANCELED_GROUP_KEY";
+    private static final String DOWNLOAD_GROUP = "DOWNLOAD_GROUP";
+    private static final String ERROR_GROUP = "ERROR_GROUP";
+    private static final String COMPLETE_GROUP = "COMPLETE_GROUP";
+    private static final String CANCEL_GROUP = "CANCEL_GROUP";
+    private static final String UPLOAD_GROUP = "UPLOAD_GROUP";
 
-    private NotificationManager mNotificationManager;
-    private Context mContext;
-    private String mServiceName;
+    private String serviceName;
+    private Context context;
+    private NotificationManager notificationManager;
 
     public NotificationUtils(Context context, String serviceName) {
-        mContext = context;
-        mServiceName = serviceName;
+        this.serviceName = serviceName;
+        this.context = context;
         init();
     }
 
     private void init() {
-        mNotificationManager = (NotificationManager) mContext.getSystemService(Context.NOTIFICATION_SERVICE);
+        notificationManager = (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
 
         // Need for Oreo, else doesn't work
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            NotificationChannel channel = new NotificationChannel(mServiceName, mContext.getString(R.string.app_name), NotificationManager.IMPORTANCE_HIGH);
+            NotificationChannel channel = new NotificationChannel(serviceName, context.getString(R.string.app_name), NotificationManager.IMPORTANCE_DEFAULT);
             channel.enableVibration(false);
-            mNotificationManager.createNotificationChannel(channel);
+            notificationManager.createNotificationChannel(channel);
         }
     }
 
-    /*
-     * Notifications
-     * */
-    private NotificationCompat.Builder getNotificationBuilder(final String title, final String info,
-                                                              final boolean isOngoing, final boolean isAlertOnce) {
-        return new NotificationCompat.Builder(App.getApp(), mServiceName)
-                .setPriority(Notification.PRIORITY_HIGH)
-                .setOngoing(isOngoing)
-                .setSmallIcon(R.drawable.ic_notify)
+    public void show(int id, Notification notification) {
+        notificationManager.notify(id, notification);
+    }
+
+    private NotificationCompat.Builder getDownloadNotificationBuilder(@NonNull String title) {
+        return new NotificationCompat.Builder(App.getApp(), serviceName)
+                .setPriority(Notification.PRIORITY_DEFAULT)
+                .setOngoing(true)
                 .setContentTitle(title)
-                .setContentText(info)
-                .setTicker(mContext.getString(R.string.app_name))
-                .setOnlyAlertOnce(isAlertOnce)
-                .setChannelId(mServiceName)
+                .setSmallIcon(R.drawable.ic_notify)
+                .setContentText(context.getString(R.string.download_manager_progress_title))
+                .setTicker(context.getString(R.string.app_name))
+                .setOnlyAlertOnce(true)
+                .setChannelId(serviceName)
+                .setGroup(DOWNLOAD_GROUP)
                 .setVisibility(NotificationCompat.VISIBILITY_PUBLIC);
     }
 
-    private NotificationCompat.Builder getDownloadNotificationBuilder(final String title, final String info) {
-        return new NotificationCompat.Builder(App.getApp(), mServiceName)
-                .setPriority(NotificationCompat.PRIORITY_DEFAULT)
-                .setVibrate(null)
+    private NotificationCompat.Builder getArchivingNotificationBuilder(@NonNull String title) {
+        return new NotificationCompat.Builder(App.getApp(), serviceName)
+                .setPriority(Notification.PRIORITY_DEFAULT)
+                .setOngoing(true)
+                .setContentTitle(title)
+                .setSmallIcon(R.drawable.ic_notify)
+                .setContentText(context.getString(R.string.download_manager_archiving_progress))
+                .setTicker(context.getString(R.string.app_name))
+                .setOnlyAlertOnce(true)
+                .setChannelId(serviceName)
+                .setGroup(DOWNLOAD_GROUP)
+                .setVisibility(NotificationCompat.VISIBILITY_PUBLIC);
+    }
+    private NotificationCompat.Builder getUploadNotificationBuilder(@NonNull String title) {
+        return new NotificationCompat.Builder(App.getApp(), serviceName)
+                .setPriority(Notification.PRIORITY_DEFAULT)
+                .setOngoing(true)
+                .setContentTitle(title)
+                .setSmallIcon(R.drawable.ic_notify)
+                .setContentText(context.getString(R.string.upload_manager_progress_title))
+                .setTicker(context.getString(R.string.app_name))
+                .setOnlyAlertOnce(true)
+                .setChannelId(serviceName)
+                .setGroup(UPLOAD_GROUP)
+                .setVisibility(NotificationCompat.VISIBILITY_PUBLIC);
+    }
+
+    public NotificationCompat.Builder getNotification(String title, @Nullable String group, @Nullable String info) {
+        return new NotificationCompat.Builder(App.getApp(), serviceName)
+                .setPriority(Notification.PRIORITY_DEFAULT)
                 .setOngoing(false)
-                .setSmallIcon(R.drawable.ic_notify)
                 .setContentTitle(title)
                 .setContentText(info)
-                .setTicker(mContext.getString(R.string.app_name))
-                .setChannelId(mServiceName)
+                .setSmallIcon(R.drawable.ic_notify)
+                .setTicker(context.getString(R.string.app_name))
+                .setOnlyAlertOnce(false)
+                .setGroup(group)
+                .setChannelId(serviceName)
                 .setVisibility(NotificationCompat.VISIBILITY_PUBLIC);
     }
 
-    private NotificationCompat.Builder getCanceledNotificationGroupBuilder() {
-        return new NotificationCompat.Builder(App.getApp(), mServiceName)
-                .setSmallIcon(R.drawable.ic_notify)
-                .setContentInfo(mContext.getString(R.string.download_manager_cancel))
-                .setChannelId(mServiceName)
-                .setGroup(CANCELED_GROUP_KEY)
-                .setGroupSummary(true);
+    public void showProgressNotification(int id, @NonNull String tag, @NonNull String title, int progress) {
+        Notification notification = getDownloadNotificationBuilder(title)
+                .addAction(R.drawable.drawable_ic_cancel_download_upload, context.getString(R.string.operation_panel_cancel_button), getDownloadIntent(tag))
+                .setProgress(100, progress, false)
+                .build();
+        notificationManager.notify(id, notification);
     }
 
-    private NotificationCompat.Builder getCanceledNotificationBuilder(final String title) {
-        return new NotificationCompat.Builder(App.getApp(), mServiceName)
-                .setSmallIcon(R.drawable.ic_notify)
-                .setContentTitle(title)
-                .setContentText(mContext.getString(R.string.download_manager_cancel))
-                .setChannelId(mServiceName)
-                .setGroup(CANCELED_GROUP_KEY);
+    public void showArchivingProgressNotification(int id, @NonNull String tag, @NonNull String title, int progress) {
+        Notification notification = getArchivingNotificationBuilder(title)
+                .addAction(R.drawable.drawable_ic_cancel_download_upload, context.getString(R.string.operation_panel_cancel_button), getDownloadIntent(tag))
+                .setProgress(100, progress, false)
+                .build();
+        notificationManager.notify(id, notification);
     }
 
-    private void showNotification(final NotificationCompat.Builder notificationBuilder, final String tag, final int id) {
-        mNotificationManager.notify(tag, id, notificationBuilder.build());
+    public void showUploadProgressNotification(int id, @NonNull String tag, @NonNull String title, int progress) {
+        Notification notification = getUploadNotificationBuilder(title)
+                .addAction(R.drawable.drawable_ic_cancel_download_upload, context.getString(R.string.operation_panel_cancel_button), getUploadIntent(tag))
+                .setProgress(100, progress, false)
+                .build();
+        notificationManager.notify(id, notification);
     }
 
-    public void showCanceledNotification(int id, String title) {
-        mNotificationManager.notify(id, getCanceledNotificationBuilder(title).build());
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-            mNotificationManager.notify(CANCELED_GROUP_ID, getCanceledNotificationGroupBuilder().build());
-        }
+    public void showErrorNotification(int id, @Nullable String title) {
+        NotificationCompat.Builder builder = getNotification(title, ERROR_GROUP, context.getString(R.string.download_manager_error));
+        notificationManager.notify(id, builder.build());
     }
 
-    public void removeNotification(final String tag, final int id) {
-        mNotificationManager.cancel(tag, id);
+    public void showUploadErrorNotification(int id, @Nullable String title) {
+        NotificationCompat.Builder builder = getNotification(title, ERROR_GROUP, context.getString(R.string.upload_manager_error));
+        notificationManager.notify(id, builder.build());
     }
 
-    public void removeNotificationById(final int id) {
-        mNotificationManager.cancel(id);
+    public void showCompleteNotification(int id, @Nullable String title, Uri uri) {
+        showDownloadedAction(id, title, uri);
     }
 
-    public void showForegroundNotification(final String id, final String title, final String info, final int total, final int progress, Service service) {
-        final NotificationCompat.Builder builder = getNotificationBuilder(title, info, true, true);
-        builder.addAction(R.drawable.drawable_ic_cancel_download_upload, mContext.getString(R.string.operation_panel_cancel_button), getUploadIntent(id))
-                .setProgress(total, progress, false);
-        service.startForeground(id.hashCode(), builder.build());
+    public void showUploadCompleteNotification(int id, @Nullable String title) {
+        showUploadedAction(id, title);
     }
 
-    public void showForegroundNotification(final String tag, final int id, final String title, final String info, final int total, final int progress, Service service) {
-        final NotificationCompat.Builder builder = getDownloadNotificationBuilder(title, info);
-        builder.setProgress(total, progress, false)
-                .addAction(R.drawable.drawable_ic_cancel_download_upload, mContext.getString(R.string.operation_panel_cancel_button), getDownloadIntent(tag));
-        service.startForeground(id, builder.build());
+    public void showCanceledNotification(int id, @Nullable String title) {
+        NotificationCompat.Builder builder = getNotification(title, CANCEL_GROUP, context.getString(R.string.download_manager_cancel));
+        notificationManager.notify(id, builder.build());
+    }
+    public void showCanceledUploadNotification(int id, @Nullable String title) {
+        NotificationCompat.Builder builder = getNotification(title, CANCEL_GROUP, context.getString(R.string.upload_manager_cancel));
+        notificationManager.notify(id, builder.build());
     }
 
-    private PendingIntent getUploadIntent(String id) {
-        Intent intent = new Intent(App.getApp(), UploadService.class);
-        intent.setAction(UploadService.ACTION_CANCEL);
-        intent.putExtra(UploadService.TAG_ID, id);
-        return PendingIntent.getService(App.getApp(), 0, intent, 0);
+    public void removeNotification(int id) {
+        notificationManager.cancel(id);
+    }
+
+    private void showDownloadedAction(final int id, final String title, final Uri uri) {
+        // Add file to default downloadFile manager
+
+        final PendingIntent contentIntent = PendingIntent.getActivity(context, 0, ActivitiesUtils.getDownloadsViewerIntent(uri), 0);
+
+        NotificationCompat.Builder builder = getNotification(title, null, context.getString(R.string.download_manager_complete))
+                .setGroupSummary(false)
+                .setCategory(NotificationCompat.CATEGORY_SERVICE)
+                .setContentIntent(contentIntent)
+                .setAutoCancel(true);
+        notificationManager.notify(id, builder.build());
+    }
+
+    private void showUploadedAction(final int id, final String title) {
+
+        Intent mainActivityIntent = new Intent(context, MainActivity.class);
+
+        final PendingIntent contentIntent = PendingIntent.getActivity(context, 0, mainActivityIntent, 0);
+
+        NotificationCompat.Builder builder = getNotification(title, null, context.getString(R.string.upload_manager_complete))
+                .setGroupSummary(false)
+                .setCategory(NotificationCompat.CATEGORY_SERVICE)
+                .setContentIntent(contentIntent);
+        notificationManager.notify(id, builder.build());
     }
 
     private PendingIntent getDownloadIntent(String id) {
-        Intent intent = new Intent(App.getApp(), DownloadService.class);
-        intent.setAction(DownloadService.ACTION_DOWNLOAD_CANCEL);
-        intent.putExtra(DownloadService.TAG_ID, id);
-        return PendingIntent.getService(App.getApp(), 0, intent, 0);
+        final Bundle bundle = new Bundle();
+        bundle.putString(DownloadReceiver.EXTRAS_KEY_ID, id);
+
+        final Intent intent = new Intent(App.getApp(), MainActivity.class);
+        intent.setAction(DownloadReceiver.DOWNLOAD_ACTION_CANCELED);
+        intent.putExtras(bundle);
+        return PendingIntent.getActivity(App.getApp(), 0, intent, PendingIntent.FLAG_UPDATE_CURRENT);
     }
 
-    public void showDownloadedNotification(final File file, final String mimeType, final String tag, final int id, final String title, final String info) {
-        // Add file to default downloadFile manager
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-            //TODO Use content resolver
-        } else {
-            final DownloadManager downloadManager = (DownloadManager) mContext.getSystemService(Context.DOWNLOAD_SERVICE);
-            if (downloadManager != null) {
-                downloadManager.addCompletedDownload(file.getName(), mContext.getString(R.string.download_manager_title), true,
-                        mimeType, file.getPath(), file.length(), false);
-            }
-        }
+    private PendingIntent getUploadIntent(String id) {
+        final Bundle bundle = new Bundle();
+        bundle.putString(UploadReceiver.EXTRAS_KEY_ID, id);
 
-
-        // On click notification - open downloadFile manager
-        final PendingIntent contentIntent = PendingIntent.getActivity(mContext, 0, ActivitiesUtils.getDownloadsViewerIntent(Uri.EMPTY), 0);
-
-        // Add action
-        final NotificationCompat.Builder builder = getNotificationBuilder(title, info, false, false);
-        builder.setCategory(NotificationCompat.CATEGORY_SERVICE);
-        builder.setContentIntent(contentIntent);
-        showNotification(builder, tag, id);
-    }
-
-    public void showInfoNotification(final String tag, final int id, final String title, final String info) {
-        final NotificationCompat.Builder builder = getNotificationBuilder(title, info, false, false);
-        builder.setCategory(NotificationCompat.CATEGORY_SERVICE);
-        showNotification(builder, tag, id);
-    }
-
-    public static long getAvailableSpaceInBytes() {
-        final StatFs statFs = new StatFs(Environment.getExternalStorageDirectory().getPath());
-        return statFs.getFreeBytes();
+        final Intent intent = new Intent(App.getApp(), MainActivity.class);
+        intent.setAction(UploadReceiver.UPLOAD_ACTION_CANCELED);
+        intent.putExtras(bundle);
+        return PendingIntent.getActivity(App.getApp(), 0, intent, PendingIntent.FLAG_UPDATE_CURRENT);
     }
 }
