+++ app_manager/appmanager/src/main/java/app/editors/manager/managers/works/DownloadWork.kt
@@ -17,7 +17,7 @@ import app.editors.manager.app.App
 import app.editors.manager.di.component.DaggerApiComponent
 import app.editors.manager.managers.receivers.DownloadReceiver
 import app.editors.manager.managers.utils.FirebaseUtils
-import app.editors.manager.managers.utils.NewNotificationUtils
+import app.editors.manager.managers.utils.NotificationUtils
 import app.editors.manager.mvp.models.explorer.Operation
 import app.editors.manager.mvp.models.request.RequestDownload
 import com.google.gson.Gson
@@ -81,7 +81,8 @@ class DownloadWork(context: Context, workerParams: WorkerParameters) : Worker(co
         }
     }
 
-    private val notificationUtils: NewNotificationUtils = NewNotificationUtils(applicationContext, TAG)
+    private val notificationUtils: NotificationUtils =
+        NotificationUtils(applicationContext, TAG)
     private var url: String? = null
     private var file: DocumentFile? = null
     private var fileName: String? = null
