+++ app_manager/appmanager/src/main/java/app/editors/manager/onedrive/managers/works/DownloadWork.kt
@@ -15,7 +15,7 @@ import app.editors.manager.app.getOneDriveServiceProvider
 import app.editors.manager.onedrive.onedrive.OneDriveResponse
 import app.editors.manager.managers.receivers.DownloadReceiver
 import app.editors.manager.managers.utils.FirebaseUtils
-import app.editors.manager.managers.utils.NewNotificationUtils
+import app.editors.manager.managers.utils.NotificationUtils
 import lib.toolkit.base.managers.utils.FileUtils
 import lib.toolkit.base.managers.utils.PathUtils
 import lib.toolkit.base.managers.utils.StringUtils
@@ -38,7 +38,8 @@ class DownloadWork(context: Context, workerParameters: WorkerParameters): Worker
         const val DOWNLOAD_ZIP_NAME = "onedrive.zip"
     }
 
-    private val notificationUtils: NewNotificationUtils = NewNotificationUtils(applicationContext, TAG)
+    private val notificationUtils: NotificationUtils =
+        NotificationUtils(applicationContext, TAG)
     private var file: DocumentFile? = null
     private var id: String? = null
     private var to: Uri? = null
