+++ app_manager/appmanager/src/main/java/app/editors/manager/ui/compose/base/BaseViews.kt
@@ -3,6 +3,7 @@ package app.editors.manager.ui.compose.base
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -11,9 +12,12 @@ import androidx.compose.material.Icon
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.material.TopAppBar
+import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.composed
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.Dp
@@ -35,10 +39,20 @@ fun CustomAppBar(@StringRes title: Int, @DrawableRes icon: Int, click: () -> Uni
                     painter = painterResource(id = icon),
                     contentDescription = "Back",
                     tint = MaterialTheme.colors.primary,
-                    modifier = Modifier.clickable { click() }
+                    modifier = Modifier.clickableRipple { click() }
                 )
             }
         },
         backgroundColor = MaterialTheme.colors.colorAppBar
+    )
+}
+
+fun Modifier.clickableRipple(onClick: () -> Unit): Modifier = composed {
+    clickable(
+        interactionSource = remember { MutableInteractionSource() },
+        indication = rememberRipple(bounded = false),
+        onClick = {
+            onClick()
+        }
     )
 }
